backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai-whisper
ffmpeg-python
openai
aiofiles
sse-starlette
torch
